Move code to 'features/users.js'
