+++ src/components/reviews/reviews.js
@@ -5,12 +5,12 @@ import Review from './review';
 import ReviewForm from './review-form';
 import styles from './reviews.module.css';
 
-import { loadUsers } from '../../redux/actions';
+import { loadUsers } from '../../redux/features/users';
 import {
   loadReviews,
   reviewsLoadedSelector,
 } from '../../redux/features/reviews';
-import { usersLoadedSelector } from '../../redux/selectors';
+import { usersLoadedSelector } from '../../redux/features/users';
 
 import Loader from '../loader';
