+++ src/redux/constants.js
@@ -1,5 +1,3 @@
-export const LOAD_USERS = 'LOAD_USERS';
-
 export const REQUEST = '_REQUEST';
 export const SUCCESS = '_SUCCESS';
 export const FAILURE = '_FAILURE';
