+++ src/redux/reducer/index.js
@@ -2,7 +2,7 @@ import order from '../features/order';
 import restaurants from '../features/restaurants';
 import products from '../features/products';
 import reviews from '../features/reviews';
-import users from './users';
+import users from '../features/users';
 
 export default {
   order,
